requirements.txt
pandas
numpy
scipy
scikit-learn
flask
joblib
pyYaml
#jupyter
#ipykernel
# -e .